[seller] refactor(245): input-field를 widgets에서 shared/ui로 이동
- features → widgets FSD 단방향 의존성 위반 해소.
