+++ apps/seller/src/features/seller-info/store-name-form/store-name-form.ui.tsx
@@ -3,7 +3,7 @@ import { ChangeEvent, useState } from 'react'
 import { Button } from '@dessert/ui'
 
 import { cn } from '@/shared/libs/utils'
-import { InputField } from '../../../widgets/input-field'
+import { InputField } from '../../../shared/ui/input-field'
 
 import {
   SELLER_INFO_CANCEL_DIALOG_CONTENT,
